feat(dashboard): split Usage into Usage / Errors tabs and rework the filters (MUL-5759)
The page answered two questions on one scroll — "what did this cost" and
"what broke" — and the second one lost. The failure breakdown sat below a
leaderboard that can run to thirty rows, and charting failures meant hiding
spend, because Errors was the fifth option of the single trend toggle.

Split by the question the reader arrives with:

- Usage tab: cost / tokens / run time / tasks KPIs, the spend trend chart,
  the agent leaderboard.
- Errors tab: failure count / rate / agents affected as real KPIs, its own
  trend chart, the failure mix, and the offender ranking.

The tab lives in `?tab=errors` because the Errors view is the half of this
page people paste at each other. The failure count rides on the tab so
"nothing broke" and "you didn't look" stay distinguishable from Usage.

Controls now say what they scope by where they live. The header carries the
two page-scoped filters; every card carries its own view switches:

- Daily / Weekly moves into each chart card, and each card offers only the
  dimensions the current range allows. That removes the hidden coupling where
  picking Weekly on a 1d range silently rewrote the range and moved every KPI
  on the page — a card-scoped control no longer reaches up to a page-scoped
  one. `handleDimChange` / `DEFAULT_DAYS_BY_DIM` are gone.
- The five permanently-expanded range segments and the bordered project Select
  become two outline Buttons, in the grammar the issues surface already uses:
  neutral while unfiltered, the filled `brand` tier plus a count once
  something is, one submenu per dimension. That is 410px of header down to
  172px, and one element family instead of two.
- The header states the viewing timezone and last refresh — every number here
  is bucketed on that timezone and nothing admitted it.

Loading and empty states are per tab, so Usage no longer waits on the two
failure rollups. All six rollups stay prefetched, so switching tabs is
instant; the per-date queries now over-fetch to the weekly span
unconditionally, which also means flipping a card between Daily and Weekly
never refetches.

dashboard-page.tsx dropped from 1550 lines to ~500; the leaderboard, the
errors view, the trend card and the filters are their own files.
