+++ packages/views/dashboard/components/dashboard-page.test.tsx
@@ -1,3 +1,4 @@
+import { useMemo, useState } from "react";
 import { describe, it, expect, afterEach, beforeEach, vi } from "vitest";
 import { cleanup, screen, within } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
@@ -244,33 +245,57 @@ vi.mock("@multica/core/runtimes/custom-pricing-store", () => {
 
 import { DashboardPage } from "./dashboard-page";
 
-// A minimal adapter — the Errors breakdown's drill-down renders an <AppLink>,
-// which reads the navigation context. Asserting on the rendered href (rather
-// than a push spy) keeps the test on the contract that matters: the row points
-// at the agent's Overview, whose ActivityTab lists its runs and each failure's
-// reason.
-const navAdapter: NavigationAdapter = {
-  push: vi.fn(),
-  replace: vi.fn(),
-  back: vi.fn(),
-  pathname: "/acme/usage",
-  searchParams: new URLSearchParams(),
-  getShareableUrl: (path: string) => `https://example.test${path}`,
-};
+const replaceSpy = vi.fn();
 
-function renderDashboard() {
-  return renderWithI18n(
-    <NavigationProvider value={navAdapter}>
+/**
+ * A navigation adapter that actually holds the query string, because the tab
+ * IS the URL: the page reads `?tab=` and writes it back through `replace`. A
+ * spy-only adapter would swallow the write and the tab could never change,
+ * which would make every Errors assertion below test a screen the user cannot
+ * reach.
+ *
+ * The adapter also backs the offender drill-down's <AppLink>. Asserting on the
+ * rendered href (rather than a push spy) keeps that test on the contract that
+ * matters: the row points at the agent's Overview, whose ActivityTab lists its
+ * runs and each failure's reason.
+ */
+function DashboardHarness({ initialSearch = "" }: { initialSearch?: string }) {
+  const [search, setSearch] = useState(initialSearch);
+  const adapter = useMemo<NavigationAdapter>(
+    () => ({
+      push: vi.fn(),
+      replace: (path: string) => {
+        replaceSpy(path);
+        setSearch(path.split("?")[1] ?? "");
+      },
+      back: vi.fn(),
+      pathname: "/acme/usage",
+      searchParams: new URLSearchParams(search),
+      getShareableUrl: (path: string) => `https://example.test${path}`,
+    }),
+    [search],
+  );
+  return (
+    <NavigationProvider value={adapter}>
       <DashboardPage />
-    </NavigationProvider>,
+    </NavigationProvider>
   );
 }
 
-// The Top offenders section — the sort control and column headers sit beside
-// the list, not inside it.
-function offenderCard(): HTMLElement {
-  return screen.getByRole("list", { name: "Top offenders" })
-    .parentElement as HTMLElement;
+function renderDashboard(initialSearch = "") {
+  return renderWithI18n(<DashboardHarness initialSearch={initialSearch} />);
+}
+
+/** Everything about failures now lives behind its own tab, so a test that
+ *  wants it has to go there first — exactly as a reader does. */
+async function openErrorsTab(user: ReturnType<typeof userEvent.setup>) {
+  await user.click(screen.getByRole("tab", { name: /Errors/ }));
+}
+
+// The offender ranking control. Named rather than reached through the DOM
+// tree so it survives the card's internal layout changing.
+function offenderSort(): HTMLElement {
+  return screen.getByRole("group", { name: "Rank offenders by" });
 }
 
 // The filled part of one offender row's bar. `role="img"` is on the fill (it
@@ -300,7 +325,7 @@ describe("DashboardPage — viewing timezone drives the query key", () => {
 
   it("uses the stored timezone in every dashboard query key", () => {
     tzRef.current = "UTC";
-    renderWithI18n(<DashboardPage />);
+    renderDashboard();
 
     const tzs = tzSegments();
     expect(tzs.length).toBeGreaterThan(0);
@@ -309,14 +334,14 @@ describe("DashboardPage — viewing timezone drives the query key", () => {
 
   it("flips the query key when the stored timezone changes", () => {
     tzRef.current = "UTC";
-    renderWithI18n(<DashboardPage />);
+    renderDashboard();
     const utcKeys = queryKeys.filter((k) => k[0] === "dashboard");
 
     queryKeys.length = 0;
     cleanup();
 
     tzRef.current = "Asia/Tokyo";
-    renderWithI18n(<DashboardPage />);
+    renderDashboard();
     const tokyoKeys = queryKeys.filter((k) => k[0] === "dashboard");
 
     expect(utcKeys.length).toBe(tokyoKeys.length);
@@ -358,24 +383,31 @@ describe("DashboardPage — failure visibility", () => {
     queryKeys.length = 0;
     dashboardDataRef.current = true;
     tzRef.current = "UTC";
+    replaceSpy.mockClear();
     cleanup();
   });
 
-  it("states the error rate with its denominator spelled out", () => {
+  it("states the error rate with its denominator spelled out", async () => {
+    const user = userEvent.setup();
+    renderDashboard();
+
+    // The Tasks tile on the Usage tab keeps its own started-tasks-only figure.
+    expect(screen.getByText("1 failed")).toBeInTheDocument();
+
+    await openErrorsTab(user);
+
     // The run-time rollup sees 1 failure out of 12 tasks — it only counts
     // tasks that actually started. The failure rollup also sees tasks that
-    // never started, so it reports 4 out of 10. The Errors card quotes the
-    // latter *with* its denominator, which is what keeps it from reading as
-    // a contradiction of the Tasks tile above.
-    renderDashboard();
-
+    // never started, so it reports 4 out of 10. The rate tile quotes the
+    // latter *with* its denominator, which is what keeps it from reading as a
+    // contradiction of the Tasks tile on the other tab.
     expect(screen.getByText("4 of 10 runs failed · 40%")).toBeInTheDocument();
-    // The Tasks tile keeps its own started-tasks-only figure.
-    expect(screen.getByText("1 failed")).toBeInTheDocument();
   });
 
-  it("breaks failures down by class and links the offending agent to its runs", () => {
+  it("breaks failures down by class and links the offending agent to its runs", async () => {
+    const user = userEvent.setup();
     renderDashboard();
+    await openErrorsTab(user);
 
     // Auth (3) outranks Timeout (1), and both are named by class rather than
     // by the raw failure_reason enum.
@@ -384,7 +416,7 @@ describe("DashboardPage — failure visibility", () => {
       "Auth3",
       "Timeout1",
     ]);
-    // The section names its own denominator: the header above it quotes a
+    // The section names its own denominator: the rate tile above quotes a
     // rate over all 10 runs, this one splits the 4 failures.
     expect(screen.getByText("Failure mix · 4")).toBeInTheDocument();
 
@@ -408,6 +440,7 @@ describe("DashboardPage — failure visibility", () => {
   it("moves the bar onto whichever metric the list is ranked by", async () => {
     const user = userEvent.setup();
     renderDashboard();
+    await openErrorsTab(user);
 
     // Agent One failed 4 of 10; the anonymous bucket failed 2 of 2. Under
     // Failures the bucket is half the leader's bar. Under Rate it is 100% —
@@ -415,7 +448,7 @@ describe("DashboardPage — failure visibility", () => {
     // exactly the mismatch this control exists to remove.
     expect(offenderBar(1).style.width).toBe("50%");
 
-    await user.click(within(offenderCard()).getByRole("button", { name: "Rate" }));
+    await user.click(within(offenderSort()).getByRole("button", { name: "Rate" }));
 
     expect(offenderBar(1).style.width).toBe("100%");
   });
@@ -423,11 +456,12 @@ describe("DashboardPage — failure visibility", () => {
   it("says which metric the list is ranked by without relying on colour", async () => {
     const user = userEvent.setup();
     renderDashboard();
+    await openErrorsTab(user);
 
     // The active option used to be a colour swap and nothing else, which is
     // invisible to a screen reader. The group is named too — "Rate, pressed"
     // means nothing until you know the group ranks the offender list.
-    const group = screen.getByRole("group", { name: "Rank offenders by" });
+    const group = offenderSort();
     expect(
       within(group).getByRole("button", { name: "Failures" }),
     ).toHaveAttribute("aria-pressed", "true");
@@ -447,8 +481,9 @@ describe("DashboardPage — failure visibility", () => {
   it("keeps a two-run agent from hijacking the rate ranking", async () => {
     const user = userEvent.setup();
     renderDashboard();
+    await openErrorsTab(user);
 
-    await user.click(within(offenderCard()).getByRole("button", { name: "Rate" }));
+    await user.click(within(offenderSort()).getByRole("button", { name: "Rate" }));
 
     // 2/2 is a 100% rate and 4/10 is 40%, but two runs is not evidence. The
     // small-sample row is demoted, not dropped — the list still has to
@@ -463,6 +498,7 @@ describe("DashboardPage — failure visibility", () => {
   it("reveals the raw failure_reason values behind the class summary", async () => {
     const user = userEvent.setup();
     renderDashboard();
+    await openErrorsTab(user);
 
     expect(
       screen.queryByText("agent_error.provider_auth_or_access"),
@@ -477,11 +513,20 @@ describe("DashboardPage — failure visibility", () => {
     ).toBeInTheDocument();
   });
 
-  it("adds Errors to the trend toggle without disturbing the other metrics", async () => {
+  it("gives failures their own chart instead of a slot on the spend toggle", async () => {
     const user = userEvent.setup();
     const { container } = renderDashboard();
 
-    await user.click(screen.getByRole("button", { name: "Errors" }));
+    // Charting failures used to mean hiding spend: Errors was the fifth option
+    // of the single trend toggle, so "what did it cost" and "what broke" could
+    // not be on screen in the same breath.
+    const metrics = within(screen.getByRole("group", { name: "Metric" }));
+    expect(
+      metrics.queryByRole("button", { name: "Errors" }),
+    ).not.toBeInTheDocument();
+    expect(metrics.getByRole("button", { name: "Tokens" })).toBeInTheDocument();
+
+    await openErrorsTab(user);
 
     expect(container).toHaveTextContent("Daily errors");
   });
@@ -492,15 +537,18 @@ describe("DashboardPage — the Errors list never exposes an agent the viewer ca
     queryKeys.length = 0;
     dashboardDataRef.current = true;
     tzRef.current = "UTC";
+    replaceSpy.mockClear();
     cleanup();
   });
 
-  it("folds an unresolvable agent into an anonymous row instead of printing its UUID", () => {
+  it("folds an unresolvable agent into an anonymous row instead of printing its UUID", async () => {
     // The failure rollups are workspace-scoped and deliberately skip
     // per-agent visibility, but the agent list does not: a private agent
     // this member can't see never appears there. Rendering `row.agentId`
     // would leak its existence, failure count and failure rate.
+    const user = userEvent.setup();
     const { container } = renderDashboard();
+    await openErrorsTab(user);
 
     expect(container).not.toHaveTextContent("0f9d1c2e-private-agent-uuid");
 
@@ -513,35 +561,78 @@ describe("DashboardPage — the Errors list never exposes an agent the viewer ca
   });
 });
 
-describe("DashboardPage — Errors card placement and density", () => {
+// The page answers two questions — "what did this cost" and "what broke" —
+// and used to answer both on one scroll, where the failure breakdown sat
+// below a leaderboard that can itself run to thirty rows (MUL-5759).
+describe("DashboardPage — the two questions are separate tabs", () => {
   beforeEach(() => {
     queryKeys.length = 0;
     dashboardDataRef.current = true;
     manyAgentsRef.current = false;
     tzRef.current = "UTC";
+    replaceSpy.mockClear();
     cleanup();
   });
 
-  it("renders the Errors card after the leaderboard, at the bottom of the page", () => {
-    // Spend is the headline; failures are the follow-up question you ask
-    // after seeing who is spending. Asserting document order rather than a
-    // class name keeps this about the reading sequence.
+  it("keeps failures off the Usage tab and reaches them in one click", async () => {
+    const user = userEvent.setup();
     renderDashboard();
 
-    // By heading, not by text: "Errors" also names the trend-chart toggle.
-    const leaderboard = screen.getByRole("heading", { name: "Leaderboard" });
-    const errors = screen.getByRole("heading", { name: "Errors" });
-
     expect(
-      leaderboard.compareDocumentPosition(errors) &
-        Node.DOCUMENT_POSITION_FOLLOWING,
-    ).toBeTruthy();
+      screen.queryByRole("list", { name: "Top offenders" }),
+    ).not.toBeInTheDocument();
+    expect(screen.getByRole("list", { name: "Leaderboard" })).toBeInTheDocument();
+
+    await openErrorsTab(user);
+
+    expect(screen.getByRole("list", { name: "Top offenders" })).toBeInTheDocument();
+    // The two per-agent rankings had near-identical shapes and used to stack
+    // one above the other; each now belongs to the question it answers.
+    expect(
+      screen.queryByRole("list", { name: "Leaderboard" }),
+    ).not.toBeInTheDocument();
+  });
+
+  it("puts the tab in the URL so an Errors view can be linked", async () => {
+    const user = userEvent.setup();
+    renderDashboard();
+
+    await openErrorsTab(user);
+    expect(replaceSpy).toHaveBeenLastCalledWith("/acme/usage?tab=errors");
+
+    // Returning to the default view drops the param instead of pinning
+    // ?tab=usage onto every link out of this page.
+    await user.click(screen.getByRole("tab", { name: "Usage" }));
+    expect(replaceSpy).toHaveBeenLastCalledWith("/acme/usage");
+  });
+
+  it("opens straight onto Errors when the URL asks for it", () => {
+    renderDashboard("tab=errors");
+
+    expect(screen.getByRole("list", { name: "Top offenders" })).toBeInTheDocument();
+  });
+
+  it("falls back to Usage for a tab value it does not recognise", () => {
+    renderDashboard("tab=nonsense");
+
+    expect(screen.getByRole("list", { name: "Leaderboard" })).toBeInTheDocument();
+  });
+
+  it("carries the failure count on the tab so a hidden view is not a silent one", () => {
+    // Without the count, "nothing broke" and "you didn't look" are the same
+    // thing from the Usage tab — which would be worse than the buried card
+    // this split replaced.
+    renderDashboard();
+
+    const errorsTab = screen.getByRole("tab", { name: /Errors/ });
+    expect(within(errorsTab).getByText("4")).toBeInTheDocument();
   });
 
   it("caps the offender list and expands it on demand", async () => {
     manyAgentsRef.current = true;
     const user = userEvent.setup();
     renderDashboard();
+    await openErrorsTab(user);
 
     const list = () => screen.getByRole("list", { name: "Top offenders" });
     // 12 agents have failures, but an unbounded list is what made this card
@@ -556,8 +647,10 @@ describe("DashboardPage — Errors card placement and density", () => {
     expect(within(list()).getAllByRole("listitem")).toHaveLength(8);
   });
 
-  it("shows no expand affordance when every offender already fits", () => {
+  it("shows no expand affordance when every offender already fits", async () => {
+    const user = userEvent.setup();
     renderDashboard();
+    await openErrorsTab(user);
 
     expect(
       screen.queryByRole("button", { name: /Show all/ }),
