+++ packages/views/locales/en/usage.json
@@ -1,10 +1,16 @@
 {
   "title": "Usage",
   "subtitle": "Token spend and agent activity across this workspace.",
+  "header": {
+    "timezone_and_updated": "{{tz}} · updated {{time}}"
+  },
   "filter": {
     "project_label": "Project",
     "all_projects": "All projects",
-    "period_label": "Period"
+    "period_label": "Period",
+    "filter_label": "Filter",
+    "active_count": "Filter · {{count}}",
+    "clear": "Clear filter"
   },
   "kpi": {
     "cost_label": "Cost · {{days}}D",
@@ -30,7 +36,6 @@
     "metric_tokens": "Tokens",
     "metric_time": "Time",
     "metric_tasks": "Tasks",
-    "metric_errors": "Errors",
     "no_data": "No usage in this window.",
     "metric_label": "Metric"
   },
@@ -43,6 +48,11 @@
     "partial_label": "{{range}} (partial · {{covered}} / 7 days)"
   },
   "errors": {
+    "kpi_failed_label": "Failed tasks · {{days}}D",
+    "kpi_failed_hint": "Of {{total}} runs",
+    "kpi_rate_label": "Failure rate · {{days}}D",
+    "kpi_agents_label": "Agents affected · {{days}}D",
+    "kpi_agents_hint": "Worst {{name}} · {{count}}",
     "tooltip_rate": "Error rate",
     "class": {
       "auth": "Auth",
