+++ packages/views/locales/ko/usage.json
@@ -1,10 +1,16 @@
 {
   "title": "사용량",
   "subtitle": "이 워크스페이스의 토큰 비용과 에이전트 활동입니다.",
+  "header": {
+    "timezone_and_updated": "{{tz}} · {{time}} 업데이트"
+  },
   "filter": {
     "project_label": "프로젝트",
     "all_projects": "모든 프로젝트",
-    "period_label": "기간"
+    "period_label": "기간",
+    "filter_label": "필터",
+    "active_count": "필터 · {{count}}",
+    "clear": "필터 지우기"
   },
   "kpi": {
     "cost_label": "비용 · {{days}}일",
@@ -30,7 +36,6 @@
     "metric_tokens": "토큰",
     "metric_time": "시간",
     "metric_tasks": "작업",
-    "metric_errors": "오류",
     "no_data": "이 기간에는 사용량이 없습니다.",
     "metric_label": "지표"
   },
@@ -43,6 +48,11 @@
     "partial_label": "{{range}} (부분 · 7일 중 {{covered}}일)"
   },
   "errors": {
+    "kpi_failed_label": "실패한 작업 · {{days}}일",
+    "kpi_failed_hint": "실행 {{total}}회 중",
+    "kpi_rate_label": "실패율 · {{days}}일",
+    "kpi_agents_label": "영향받은 에이전트 · {{days}}일",
+    "kpi_agents_hint": "최다 {{name}} · {{count}}회",
     "tooltip_rate": "오류율",
     "class": {
       "auth": "인증",
